requirements.txt
Flask==1.1.2
Flask-WTF==0.14.3
python-dateutil==2.8.1
PyYAML==5.3.1